✨ feat: boketeNominal + iiwakeBattleNominal parity fixtures (S3a item 3)
First cross-language witnesses of `assign target: all` (`shared_assignment`)
and `speak_each`, and the first fixtures carrying a non-empty `extraData`
across the Swift-JSON → Kotlin `AnyCodableValue` boundary — gated by a
dedicated decode test before any transcript comparison. Both green on both
rungs with an empty ledger. `unreachableClasses` reason strings for
`SCOREBOARD_ORDERING` / `LINT_PREDICATE_DIVERGENCE` re-verified against
the six-fixture set (kanji/hiragana names are still BMP; NFC-only keys).

The fixture roster moves to `ParityFixtureEmitter+Specs.swift` and the
per-phase sweep to `ParityFixtureEmitterTests+Exercise.swift` — both files
hit SwiftLint's length caps as fixtures accumulate; no behaviour change.

Part of #1605
