+++ shared/engine/src/commonTest/kotlin/com/pastura/engine/ParityGolden.kt
@@ -595,6 +595,252 @@ internal object ParityGolden {
     )
 
     /**
+     * Happy path, and the first witness of `assign` with `target: all` — the round-shared assignment that reaches the transcript as a `shared_assignment` event rather than the per-agent `assignment` word wolf emits. It is RNG-free for the same reason the other nominal fixtures are: the topic is picked by round index, not by draw, so two rounds walk the two `topics:` entries deterministically.
+     *
+     * **It is also the first fixture whose scenario carries a non-empty `extraData`.** `topics:` crosses the Swift-JSON -> Kotlin boundary through two independently hand-written codecs — Swift's `AnyCodableValue.encode(to:)` single-value container and Kotlin's `JsonElement`-inspecting `AnyCodableValueSerializer` — which every earlier fixture left unexercised by crossing an empty map. `someGoldenScenarioCarriesAnArrayValuedExtraDataEntry` gates that crossing separately, before the transcripts are compared at all.
+     *
+     * The `speak_all` / `vote` / `vote_tally` / `summarize` spine it runs is already witnessed by `targetScoreRaceNominal`, so what this fixture adds is the assignment head and the payload crossing, not a second copy of the voting half.
+     */
+    internal val boketeNominal: Fixture = Fixture(
+        name = "boketeNominal",
+        scenarioJson = """
+{"agentCount":5,"context":"あなたはお笑い大喜利バトル番組の出場者です。 写真のお題が出されるので、一言ボケをつけてください。 その後、他の参加者のボケを見て、自分以外で一番面白いものに投票します。 あなたのキャラクターのお笑いスタイルを活かしてください。\n","description":"5人の芸人がお題の写真に一言ボケをつけ、互いに投票する大喜利バトル。 各自のお笑いスタイルを活かして最高のボケを競う。\n","extraData":{"topics":["スーツを着た猫が真剣な顔で会議室のホワイトボードの前に立っている写真","おばあちゃんがスケボーで坂道を爆走している写真"]},"id":"bokete","language":"ja","name":"ボケて","personas":[{"description":"【立場】切れ味鋭いツッコミ師 【目的】短く鋭い一言で笑いを取る。 例:「お前が言うな」「それ逆だろ」のような切れ味のある一言。\n","name":"ツッコミ太郎"},{"description":"【立場】シュール芸人 【目的】意味不明な方向に飛躍する。脈絡がないほど良い。 例:「3年後にこれの意味がわかる」「宇宙の裏側ではこれが普通」\n","name":"シュール子"},{"description":"【立場】あるあるネタの達人 【目的】日常の共感ポイントに落とし込む。 例:「月曜の朝、みんなこうなってる」「上司がいない日の職場」\n","name":"あるある先生"},{"description":"【立場】毒舌ブラックユーモア担当 【目的】皮肉の効いた少し毒のあるボケ。 例:「給料日前日の財布の中身」「転職サイトの閲覧履歴」\n","name":"ブラック山田"},{"description":"【立場】素で面白い天然ボケ 【目的】本気で的外れなことを言う。計算ではなく素で面白い。 例:「え、これ火曜日ですか？」「あ、おいしそう」\n","name":"天然ボケ美"}],"phases":[{"source":"topics","target":"all","type":"assign"},{"outputSchema":{"inner_thought":"string","statement":"string"},"prompt":"お題: {assigned_topic}\nこのお題の写真に一言ボケをつけてください。 あなたのお笑いスタイルを活かして、面白い一言を考えてください。\n","type":"speak_all"},{"excludeSelf":true,"outputSchema":{"reason":"string","vote":"string"},"prompt":"お題: {assigned_topic}\n他の参加者のボケ: {conversation_log}\n自分以外で一番面白いと思うボケに投票してください。 投票先は必ず以下の名前のいずれかを正確に書いてください: {candidates}\n","type":"vote"},{"logic":"vote_tally","type":"score_calc"},{"template":"ラウンド{current_round}結果: {scoreboard}\n","type":"summarize"}],"rounds":2}
+""".trimIndent(),
+        responses = listOf(
+            """{"statement": "statement 0", "inner_thought": "inner_thought 0"}""",
+            """{"statement": "statement 1", "inner_thought": "inner_thought 1"}""",
+            """{"statement": "statement 2", "inner_thought": "inner_thought 2"}""",
+            """{"statement": "statement 3", "inner_thought": "inner_thought 3"}""",
+            """{"statement": "statement 4", "inner_thought": "inner_thought 4"}""",
+            """{"vote": "シュール子", "reason": "reason 5"}""",
+            """{"vote": "あるある先生", "reason": "reason 6"}""",
+            """{"vote": "ブラック山田", "reason": "reason 7"}""",
+            """{"vote": "天然ボケ美", "reason": "reason 8"}""",
+            """{"vote": "ツッコミ太郎", "reason": "reason 9"}""",
+            """{"statement": "statement 10", "inner_thought": "inner_thought 10"}""",
+            """{"statement": "statement 11", "inner_thought": "inner_thought 11"}""",
+            """{"statement": "statement 12", "inner_thought": "inner_thought 12"}""",
+            """{"statement": "statement 13", "inner_thought": "inner_thought 13"}""",
+            """{"statement": "statement 14", "inner_thought": "inner_thought 14"}""",
+            """{"vote": "シュール子", "reason": "reason 15"}""",
+            """{"vote": "あるある先生", "reason": "reason 16"}""",
+            """{"vote": "ブラック山田", "reason": "reason 17"}""",
+            """{"vote": "天然ボケ美", "reason": "reason 18"}""",
+            """{"vote": "ツッコミ太郎", "reason": "reason 19"}""",
+        ),
+        transcript = listOf(
+            """{"attempt":0,"event":"round_started","round":1,"t":0,"total_rounds":2,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"shared_assignment","t":0,"type":"event","value":"スーツを着た猫が真剣な顔で会議室のホワイトボードの前に立っている写真"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[1],"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 0","statement":"statement 0"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 1","statement":"statement 1"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 2","statement":"statement 2"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 3","statement":"statement 3"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 4","statement":"statement 4"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[1],"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"agent_output","fields":{"reason":"reason 5","vote":"シュール子"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"agent_output","fields":{"reason":"reason 6","vote":"あるある先生"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"agent_output","fields":{"reason":"reason 7","vote":"ブラック山田"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"agent_output","fields":{"reason":"reason 8","vote":"天然ボケ美"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"agent_output","fields":{"reason":"reason 9","vote":"ツッコミ太郎"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"vote_results","t":0,"tallies":{"あるある先生":1,"シュール子":1,"ツッコミ太郎":1,"ブラック山田":1,"天然ボケ美":1},"type":"event","votes":{"あるある先生":"ブラック山田","シュール子":"あるある先生","ツッコミ太郎":"シュール子","ブラック山田":"天然ボケ美","天然ボケ美":"ツッコミ太郎"}}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"score_update","scores":{"あるある先生":1,"シュール子":1,"ツッコミ太郎":1,"ブラック山田":1,"天然ボケ美":1},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"summary","t":0,"type":"event","value":"ラウンド1結果: {\"あるある先生\": 1, \"シュール子\": 1, \"ツッコミ太郎\": 1, \"ブラック山田\": 1, \"天然ボケ美\": 1}\n"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"round_completed","round":1,"scores":{"あるある先生":1,"シュール子":1,"ツッコミ太郎":1,"ブラック山田":1,"天然ボケ美":1},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"round_started","round":2,"t":0,"total_rounds":2,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"shared_assignment","t":0,"type":"event","value":"おばあちゃんがスケボーで坂道を爆走している写真"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[1],"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 10","statement":"statement 10"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 11","statement":"statement 11"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 12","statement":"statement 12"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 13","statement":"statement 13"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 14","statement":"statement 14"},"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[1],"phase_type":"speak_all","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ツッコミ太郎","attempt":0,"event":"agent_output","fields":{"reason":"reason 15","vote":"シュール子"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"シュール子","attempt":0,"event":"agent_output","fields":{"reason":"reason 16","vote":"あるある先生"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"あるある先生","attempt":0,"event":"agent_output","fields":{"reason":"reason 17","vote":"ブラック山田"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"ブラック山田","attempt":0,"event":"agent_output","fields":{"reason":"reason 18","vote":"天然ボケ美"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"天然ボケ美","attempt":0,"event":"agent_output","fields":{"reason":"reason 19","vote":"ツッコミ太郎"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"vote_results","t":0,"tallies":{"あるある先生":1,"シュール子":1,"ツッコミ太郎":1,"ブラック山田":1,"天然ボケ美":1},"type":"event","votes":{"あるある先生":"ブラック山田","シュール子":"あるある先生","ツッコミ太郎":"シュール子","ブラック山田":"天然ボケ美","天然ボケ美":"ツッコミ太郎"}}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"score_update","scores":{"あるある先生":2,"シュール子":2,"ツッコミ太郎":2,"ブラック山田":2,"天然ボケ美":2},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"summary","t":0,"type":"event","value":"ラウンド2結果: {\"あるある先生\": 2, \"シュール子\": 2, \"ツッコミ太郎\": 2, \"ブラック山田\": 2, \"天然ボケ美\": 2}\n"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"round_completed","round":2,"scores":{"あるある先生":2,"シュール子":2,"ツッコミ太郎":2,"ブラック山田":2,"天然ボケ美":2},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"simulation_completed","t":0,"type":"event"}""",
+        ),
+        callCount = 20,
+    )
+
+    /**
+     * Happy path, and the first witness of `speak_each` — sequential turns where each agent sees the conversation so far, against the simultaneous `speak_all` every other fixture runs. Paired with `boketeNominal`, which shares the same `assign(target: all)` / `vote` / `vote_tally` / `summarize` spine and differs only in that one phase, so a divergence appearing here and not there localizes to the speech handler rather than to the scenario.
+     *
+     * **Why a `DemoPresets/` scenario is admitted.** The parity fixtures otherwise draw from `Resources/Presets/`, and reaching outside that directory is a deliberate exception rather than an oversight: this is shipped YAML (`BundledPresetResolver` resolves it, and `check_demo_replay_drift.py` hashes its bytes), and it is the only RNG-free bundled preset that runs `speak_each` at all. The alternative was a hand-authored fixture under `tools/harness/Fixtures/`, which would witness `speak_each` against a scenario no user ever runs.
+     */
+    internal val iiwakeBattleNominal: Fixture = Fixture(
+        name = "iiwakeBattleNominal",
+        scenarioJson = """
+{"agentCount":4,"context":"あなたは言い訳バトル番組「アリバイの達人」の出場者です。 全員が同じ「やらかし」を犯した設定で、順番に言い訳を発表します。 すでに出た言い訳とかぶってはいけません。前の人より印象に残る言い訳を しないと埋もれます。ただし話を盛りすぎて嘘くさくなったら逆効果です。 全員の発表が終わったら、自分以外で一番見事な言い訳に投票します。\n","description":"やらかしたシチュエーションに対して順番に言い訳をしていくバトル。 前の人の言い訳より大胆でスケールの大きい言い訳を重ねないと埋もれる、 後攻有利と自爆リスクが同居した順番制の大喜利。\n","extraData":{"topics":["やらかし「大事な会議に2時間遅刻した」","やらかし「同僚の机に置いてあったプリンを食べてしまった」"]},"id":"iiwake_battle_v1","language":"ja","name":"言い訳エスカレーション","personas":[{"description":"【立場】何でも国際問題のせいにするスケール詐欺師 【目的】時差・為替・外交問題など世界規模の理由に責任転嫁して笑いを取る。 例:「サマータイムの廃止が決まった影響で」「円安がここまでとは」\n","name":"国際派ジョージ"},{"description":"【立場】疑似科学で武装する理系風言い訳師 【目的】もっともらしい専門用語で煙に巻く言い訳で笑いを取る。 例:「量子的に観測するまで遅刻は確定しない」「腸内細菌の総意です」\n","name":"科学者リケ美"},{"description":"【立場】人情に訴える涙の言い訳師 【目的】聞いた人が思わず許したくなる、嘘くさいのに泣ける人情話で笑いを取る。 例:「捨て猫が、どうしても離れてくれなくて」「亡き祖母との約束が」\n","name":"浪花節たけし"},{"description":"【立場】言い訳を放棄して堂々と開き直る確信犯 【目的】謝るどころか逆に相手を説得し始める図々しさで笑いを取る。 例:「むしろ感謝してほしい」「これは遅刻ではなく様式美」\n","name":"開き直りマコ"}],"phases":[{"source":"topics","target":"all","type":"assign"},{"outputSchema":{"inner_thought":"string","statement":"string"},"prompt":"{assigned_topic}\nこれまでに出た言い訳: {conversation_log}\nあなたの番です。すでに出た言い訳とかぶらないように、 あなたの芸風で前の人より印象に残る言い訳を一言で発表してください。\n","type":"speak_each"},{"excludeSelf":true,"outputSchema":{"reason":"string","vote":"string"},"prompt":"{assigned_topic}\n全員の言い訳: {conversation_log}\n自分以外で一番見事だと思う言い訳に投票してください。 投票先は必ず以下の名前のいずれかを正確に書いてください: {candidates}\n","type":"vote"},{"logic":"vote_tally","type":"score_calc"},{"template":"ラウンド{current_round}結果: {scoreboard}\n","type":"summarize"}],"rounds":2}
+""".trimIndent(),
+        responses = listOf(
+            """{"statement": "statement 0", "inner_thought": "inner_thought 0"}""",
+            """{"statement": "statement 1", "inner_thought": "inner_thought 1"}""",
+            """{"statement": "statement 2", "inner_thought": "inner_thought 2"}""",
+            """{"statement": "statement 3", "inner_thought": "inner_thought 3"}""",
+            """{"vote": "科学者リケ美", "reason": "reason 4"}""",
+            """{"vote": "浪花節たけし", "reason": "reason 5"}""",
+            """{"vote": "開き直りマコ", "reason": "reason 6"}""",
+            """{"vote": "国際派ジョージ", "reason": "reason 7"}""",
+            """{"statement": "statement 8", "inner_thought": "inner_thought 8"}""",
+            """{"statement": "statement 9", "inner_thought": "inner_thought 9"}""",
+            """{"statement": "statement 10", "inner_thought": "inner_thought 10"}""",
+            """{"statement": "statement 11", "inner_thought": "inner_thought 11"}""",
+            """{"vote": "科学者リケ美", "reason": "reason 12"}""",
+            """{"vote": "浪花節たけし", "reason": "reason 13"}""",
+            """{"vote": "開き直りマコ", "reason": "reason 14"}""",
+            """{"vote": "国際派ジョージ", "reason": "reason 15"}""",
+        ),
+        transcript = listOf(
+            """{"attempt":0,"event":"round_started","round":1,"t":0,"total_rounds":2,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"shared_assignment","t":0,"type":"event","value":"やらかし「大事な会議に2時間遅刻した」"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[1],"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 0","statement":"statement 0"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 1","statement":"statement 1"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 2","statement":"statement 2"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 3","statement":"statement 3"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[1],"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"agent_output","fields":{"reason":"reason 4","vote":"科学者リケ美"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"agent_output","fields":{"reason":"reason 5","vote":"浪花節たけし"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"agent_output","fields":{"reason":"reason 6","vote":"開き直りマコ"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"agent_output","fields":{"reason":"reason 7","vote":"国際派ジョージ"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"vote_results","t":0,"tallies":{"国際派ジョージ":1,"浪花節たけし":1,"科学者リケ美":1,"開き直りマコ":1},"type":"event","votes":{"国際派ジョージ":"科学者リケ美","浪花節たけし":"開き直りマコ","科学者リケ美":"浪花節たけし","開き直りマコ":"国際派ジョージ"}}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"score_update","scores":{"国際派ジョージ":1,"浪花節たけし":1,"科学者リケ美":1,"開き直りマコ":1},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"summary","t":0,"type":"event","value":"ラウンド1結果: {\"国際派ジョージ\": 1, \"浪花節たけし\": 1, \"科学者リケ美\": 1, \"開き直りマコ\": 1}\n"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"round_completed","round":1,"scores":{"国際派ジョージ":1,"浪花節たけし":1,"科学者リケ美":1,"開き直りマコ":1},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"round_started","round":2,"t":0,"total_rounds":2,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"shared_assignment","t":0,"type":"event","value":"やらかし「同僚の机に置いてあったプリンを食べてしまった」"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[0],"phase_type":"assign","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[1],"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 8","statement":"statement 8"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 9","statement":"statement 9"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 10","statement":"statement 10"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"agent_output","fields":{"inner_thought":"inner_thought 11","statement":"statement 11"},"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[1],"phase_type":"speak_each","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"国際派ジョージ","attempt":0,"event":"agent_output","fields":{"reason":"reason 12","vote":"科学者リケ美"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"科学者リケ美","attempt":0,"event":"agent_output","fields":{"reason":"reason 13","vote":"浪花節たけし"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"浪花節たけし","attempt":0,"event":"agent_output","fields":{"reason":"reason 14","vote":"開き直りマコ"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"inference_started","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"duration_seconds":0,"event":"inference_completed","t":0,"type":"event"}""",
+            """{"agent":"開き直りマコ","attempt":0,"event":"agent_output","fields":{"reason":"reason 15","vote":"国際派ジョージ"},"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"vote_results","t":0,"tallies":{"国際派ジョージ":1,"浪花節たけし":1,"科学者リケ美":1,"開き直りマコ":1},"type":"event","votes":{"国際派ジョージ":"科学者リケ美","浪花節たけし":"開き直りマコ","科学者リケ美":"浪花節たけし","開き直りマコ":"国際派ジョージ"}}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[2],"phase_type":"vote","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"score_update","scores":{"国際派ジョージ":2,"浪花節たけし":2,"科学者リケ美":2,"開き直りマコ":2},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[3],"phase_type":"score_calc","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"phase_started","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"summary","t":0,"type":"event","value":"ラウンド2結果: {\"国際派ジョージ\": 2, \"浪花節たけし\": 2, \"科学者リケ美\": 2, \"開き直りマコ\": 2}\n"}""",
+            """{"attempt":0,"event":"phase_completed","phase_path":[4],"phase_type":"summarize","t":0,"type":"event"}""",
+            """{"attempt":0,"event":"round_completed","round":2,"scores":{"国際派ジョージ":2,"浪花節たけし":2,"科学者リケ美":2,"開き直りマコ":2},"t":0,"type":"event"}""",
+            """{"attempt":0,"event":"simulation_completed","t":0,"type":"event"}""",
+        ),
+        callCount = 16,
+    )
+
+    /**
      * Structural negative control. The sibling divergent fixture drives a VALUE divergence only; this one exists so the ledger's kind-coverage guard has a `Structural` entry to hold, and so a divergence class and its entries cannot be deleted together unnoticed — the way ADR-021 § Amendment 2026-08-06 retired `SCHEMA_GUARD_POSITION` and silently cost the control its only structural arm.
      *
      * Call 1 drives it. Swift's schema-guarded multi-object salvage (#907) accepts the first object when every expected key is present and non-empty, returning an `agentOutput` after ONE backend call. Kotlin's `extractFirstJsonObject` returns object-like residue unchanged, so the parse fails and the turn exhausts its retry budget into a `turnSkipped` after THREE. Paired parser tests fed byte-identical input pin both behaviours, so neither can drift silently.
@@ -639,6 +885,8 @@ internal object ParityGolden {
         targetScoreRaceNominal,
         targetScoreRaceDivergent,
         prisonersDilemmaNominal,
+        boketeNominal,
+        iiwakeBattleNominal,
         parityStructuralControl,
     )
 }
